reafcor: news layout
