+++ GoodNews.Client/src/Components/News/ArticleCard/index.jsx
@@ -1,0 +1,52 @@
+import React from 'react';
+import {makeStyles} from '@material-ui/core/styles';
+import Card from '@material-ui/core/Card';
+import CardActionArea from '@material-ui/core/CardActionArea';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
+import Grid from "@material-ui/core/Grid";
+import {NavLink} from "react-router-dom";
+
+const useStyles = makeStyles({
+    card: {
+        maxWidth: 345,
+    },
+    media: {
+        height: 140,
+    },
+});
+
+const ArticleCard = (props) => {
+    const classes = useStyles();
+
+    return (
+        <Grid item >
+            <Card className={classes.card}>
+                <CardActionArea>
+                    <CardMedia
+                        className={classes.media}
+                        image={props.article.thumbnailUrl}
+                    />
+                    <CardContent>
+                        <Typography gutterBottom variant="h6" component="h6">
+                            {props.article.title}
+                        </Typography>
+                        <Typography variant="body2" color="textSecondary" component="p" >
+                            {props.article.description}
+                        </Typography>
+                    </CardContent>
+                </CardActionArea>
+                <CardActions>
+                    <Button variant="contained" color="primary" size="small" href={'news/'+ props.article.id}>
+                        Подробнее
+                    </Button>
+                </CardActions>
+            </Card>
+        </Grid>
+    );
+}
+
+export default ArticleCard;
