+++ GoodNews.Client/src/Components/News/ArticleDetails/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import s from './style.module.scss'
+
+const ArticleDetails = (props) => {
+    return (
+        <div>
+            ArticleDetails
+        </div>
+    )
+}
+ export default ArticleDetails;
