+++ GoodNews.Client/src/Components/News/style.module.scss
@@ -1,0 +1,3 @@
+.newsContainer {
+  margin-top: 20px;
+}
